Move auth.py file into magiclinkbackend app
